# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Word Dictation API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Textbook {
  id: number;
  name: string;
  /** @nullable */
  description: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateTextbookBody {
  name: string;
  /** @nullable */
  description?: string | null;
}

export interface UpdateTextbookBody {
  name?: string;
  /** @nullable */
  description?: string | null;
}

export interface Unit {
  id: number;
  textbookId: number;
  name: string;
  orderIndex: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateUnitBody {
  name: string;
  orderIndex: number;
}

export interface UpdateUnitBody {
  name?: string;
  orderIndex?: number;
}

export interface Word {
  id: number;
  unitId: number;
  english: string;
  chinese: string;
  /** @nullable */
  phonetic: string | null;
  /** @nullable */
  partOfSpeech: string | null;
  orderIndex: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateWordBody {
  english: string;
  chinese: string;
  /** @nullable */
  phonetic?: string | null;
  /** @nullable */
  partOfSpeech?: string | null;
  orderIndex: number;
}

export interface UpdateWordBody {
  english?: string;
  chinese?: string;
  /** @nullable */
  phonetic?: string | null;
  /** @nullable */
  partOfSpeech?: string | null;
  orderIndex?: number;
}

export interface ImportWordsBody {
  unitId: number;
  words: CreateWordBody[];
}

export interface ImportResult {
  imported: number;
  total: number;
}

export interface ExportData {
  unitName: string;
  textbookName: string;
  words: Word[];
}

export type GenerateDictationBodyDirection =
  (typeof GenerateDictationBodyDirection)[keyof typeof GenerateDictationBodyDirection];

export const GenerateDictationBodyDirection = {
  zh_to_en: "zh_to_en",
  en_to_zh: "en_to_zh",
} as const;

export type GenerateDictationBodyStyle =
  (typeof GenerateDictationBodyStyle)[keyof typeof GenerateDictationBodyStyle];

export const GenerateDictationBodyStyle = {
  underline: "underline",
  table: "table",
} as const;

export interface GenerateDictationBody {
  unitIds: number[];
  direction: GenerateDictationBodyDirection;
  style: GenerateDictationBodyStyle;
  shuffle: boolean;
  showPhonetic: boolean;
  title?: string;
  columns: number;
}

export interface DictationItem {
  prompt: string;
  answer: string;
  /** @nullable */
  phonetic: string | null;
  /** @nullable */
  partOfSpeech: string | null;
}

export interface DictationSheet {
  title: string;
  direction: string;
  style: string;
  items: DictationItem[];
  generatedAt: string;
}

export interface StatsSummary {
  totalTextbooks: number;
  totalUnits: number;
  totalWords: number;
}