# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Word Dictation API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all textbooks
 */
export const ListTextbooksResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().nullable(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListTextbooksResponse = zod.array(ListTextbooksResponseItem);

/**
 * @summary Create a textbook
 */
export const CreateTextbookBody = zod.object({
  name: zod.string(),
  description: zod.string().nullish(),
});

/**
 * @summary Get a textbook by ID
 */
export const GetTextbookParams = zod.object({
  id: zod.coerce.number(),
});

export const GetTextbookResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().nullable(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Update a textbook
 */
export const UpdateTextbookParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateTextbookBody = zod.object({
  name: zod.string().optional(),
  description: zod.string().nullish(),
});

export const UpdateTextbookResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().nullable(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Delete a textbook
 */
export const DeleteTextbookParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List units for a textbook
 */
export const ListUnitsParams = zod.object({
  textbookId: zod.coerce.number(),
});

export const ListUnitsResponseItem = zod.object({
  id: zod.number(),
  textbookId: zod.number(),
  name: zod.string(),
  orderIndex: zod.number(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListUnitsResponse = zod.array(ListUnitsResponseItem);

/**
 * @summary Create a unit
 */
export const CreateUnitParams = zod.object({
  textbookId: zod.coerce.number(),
});

export const CreateUnitBody = zod.object({
  name: zod.string(),
  orderIndex: zod.number(),
});

/**
 * @summary Update a unit
 */
export const UpdateUnitParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateUnitBody = zod.object({
  name: zod.string().optional(),
  orderIndex: zod.number().optional(),
});

export const UpdateUnitResponse = zod.object({
  id: zod.number(),
  textbookId: zod.number(),
  name: zod.string(),
  orderIndex: zod.number(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Delete a unit
 */
export const DeleteUnitParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List words for a unit
 */
export const ListWordsParams = zod.object({
  unitId: zod.coerce.number(),
});

export const ListWordsResponseItem = zod.object({
  id: zod.number(),
  unitId: zod.number(),
  english: zod.string(),
  chinese: zod.string(),
  phonetic: zod.string().nullable(),
  partOfSpeech: zod.string().nullable(),
  orderIndex: zod.number(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListWordsResponse = zod.array(ListWordsResponseItem);

/**
 * @summary Create a word
 */
export const CreateWordParams = zod.object({
  unitId: zod.coerce.number(),
});

export const CreateWordBody = zod.object({
  english: zod.string(),
  chinese: zod.string(),
  phonetic: zod.string().nullish(),
  partOfSpeech: zod.string().nullish(),
  orderIndex: zod.number(),
});

/**
 * @summary Update a word
 */
export const UpdateWordParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateWordBody = zod.object({
  english: zod.string().optional(),
  chinese: zod.string().optional(),
  phonetic: zod.string().nullish(),
  partOfSpeech: zod.string().nullish(),
  orderIndex: zod.number().optional(),
});

export const UpdateWordResponse = zod.object({
  id: zod.number(),
  unitId: zod.number(),
  english: zod.string(),
  chinese: zod.string(),
  phonetic: zod.string().nullable(),
  partOfSpeech: zod.string().nullable(),
  orderIndex: zod.number(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Delete a word
 */
export const DeleteWordParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Import words from Excel data
 */
export const ImportWordsBody = zod.object({
  unitId: zod.number(),
  words: zod.array(
    zod.object({
      english: zod.string(),
      chinese: zod.string(),
      phonetic: zod.string().nullish(),
      partOfSpeech: zod.string().nullish(),
      orderIndex: zod.number(),
    }),
  ),
});

export const ImportWordsResponse = zod.object({
  imported: zod.number(),
  total: zod.number(),
});

/**
 * @summary Export words as JSON for Excel download
 */
export const ExportWordsParams = zod.object({
  unitId: zod.coerce.number(),
});

export const ExportWordsResponse = zod.object({
  unitName: zod.string(),
  textbookName: zod.string(),
  words: zod.array(
    zod.object({
      id: zod.number(),
      unitId: zod.number(),
      english: zod.string(),
      chinese: zod.string(),
      phonetic: zod.string().nullable(),
      partOfSpeech: zod.string().nullable(),
      orderIndex: zod.number(),
      createdAt: zod.string(),
      updatedAt: zod.string(),
    }),
  ),
});

/**
 * @summary Generate dictation sheet preview data
 */
export const GenerateDictationPreviewBody = zod.object({
  unitIds: zod.array(zod.number()),
  direction: zod.enum(["zh_to_en", "en_to_zh"]),
  style: zod.enum(["underline", "table"]),
  shuffle: zod.boolean(),
  showPhonetic: zod.boolean(),
  title: zod.string().optional(),
  columns: zod.number(),
});

export const GenerateDictationPreviewResponse = zod.object({
  title: zod.string(),
  direction: zod.string(),
  style: zod.string(),
  items: zod.array(
    zod.object({
      prompt: zod.string(),
      answer: zod.string(),
      phonetic: zod.string().nullable(),
      partOfSpeech: zod.string().nullable(),
    }),
  ),
  generatedAt: zod.string(),
});

/**
 * @summary Get overall stats summary
 */
export const GetStatsSummaryResponse = zod.object({
  totalTextbooks: zod.number(),
  totalUnits: zod.number(),
  totalWords: zod.number(),
});